remove extra space below moctale meter & chnage to show only integer part of review percentage

+++ app/src/main/java/com/manishjajoriya/moctale/presentation/contentScreen/component/MoctaleMeter.kt
@@ -97,7 +97,7 @@ fun MoctaleMeter(
                 override fun onValueSelected(e: Entry?, h: Highlight?) {
                   if (e is PieEntry) {
                     val type = e.label
-                    val percentage = e.value
+                    val percentage = e.value.toInt()
 
                     val (color, votes) =
                         when (type) {
